Move storage event manager to its own file
